Delete RenderItem component,use RenderAlertItem as RenderItem
