+++ components/RenderItem/RenderItem.tsx
@@ -1,42 +1,67 @@
-import { COLORS, Div, Divider, Text } from "cherag-ui";
-import { FC } from "react";
+import { COLORS, Div, Divider, SIZES, Text } from "cherag-ui";
+import { FC, ReactNode } from "react";
 import SyntaxHighlighter from "react-syntax-highlighter";
 import { docco } from "react-syntax-highlighter/dist/esm/styles/hljs";
 import { ProjectColor } from "../../theme";
-import { RenderItemProps } from "./types";
+import { ProperticsProps, UsecaseArayObjProps } from "./types";
+
+export type RenderItemProps = {
+  componentName: string;
+  componentDescription: string;
+  importCodeString: string;
+  usecaseCodeStringArr: UsecaseArayObjProps[];
+  children: ReactNode;
+  propertics: ProperticsProps[];
+};
 
 export const RenderItem: FC<RenderItemProps> = ({
   componentName,
+  componentDescription,
   importCodeString,
-  useCaseCodeString,
+  usecaseCodeStringArr,
   children,
   propertics,
 }) => {
   return (
     <Div style={{ marginTop: 80, marginLeft: 80, height: "auto" }}>
-      <Text style={{ fontSize: 30, fontWeight: "bold", marginBottom: 20 }}>
+      <Text style={{ fontSize: 30, fontWeight: "bold", paddingBottom: 30 }}>
         {componentName}
       </Text>
-      <Text style={{ marginBottom: 25 }}>
-        Component for displaying messages, notifications, or other application
-        state.
-      </Text>
+      <Text style={{ paddingBottom: 25 }}>{componentDescription}</Text>
 
-      <SyntaxHighlighter language="tsx" style={docco}>
-        {importCodeString}
-      </SyntaxHighlighter>
-      <Text style={{ fontSize: 30, fontWeight: "bold", marginVertical: 20 }}>
+      <Div style={{}}>
+        <SyntaxHighlighter language="tsx" style={docco}>
+          {importCodeString}
+        </SyntaxHighlighter>
+      </Div>
+      <Text
+        style={{
+          fontSize: 30,
+          fontWeight: "bold",
+          marginTop: 25,
+        }}
+      >
         Usage
       </Text>
-      <SyntaxHighlighter language="tsx" style={docco}>
-        {useCaseCodeString}
-      </SyntaxHighlighter>
+      {usecaseCodeStringArr.map((item) => {
+        return (
+          <Div key={item.name}>
+            <Text style={{ fontSize: 24, marginVertical: 20 }}>
+              {item.name}
+            </Text>
+            <SyntaxHighlighter language="tsx" style={docco}>
+              {item.code}
+            </SyntaxHighlighter>
+          </Div>
+        );
+      })}
+
       <Div
         style={{
           marginTop: 30,
           backgroundColor: COLORS.lightGray5,
           width: "100%",
-          height: 200,
+          height: "30%",
           alignItems: "center",
           justifyContent: "center",
         }}
@@ -48,8 +73,8 @@ export const RenderItem: FC<RenderItemProps> = ({
         Props
       </Text>
       <Div style={{ flexDirection: "row" }}>
-        <Text style={{ fontWeight: "bold", width: 100 }}>Name</Text>
-        <Text style={{ fontWeight: "bold", width: 100 }}>Type</Text>
+        <Text style={{ fontWeight: "bold", width: 200 }}>Name</Text>
+        <Text style={{ fontWeight: "bold", width: 200 }}>Type</Text>
         <Text style={{ fontWeight: "bold" }}>Description</Text>
       </Div>
       <Divider
@@ -61,10 +86,10 @@ export const RenderItem: FC<RenderItemProps> = ({
       />
       {propertics.map((item) => {
         return (
-          <>
+          <Div key={item.name}>
             <Div key={item.name} style={{ flexDirection: "row" }}>
-              <Text style={{ width: 100 }}>{item.name}</Text>
-              <Text style={{ width: 100 }}>{item.type}</Text>
+              <Text style={{ width: 200 }}>{item.name}</Text>
+              <Text style={{ width: 200 }}>{item.type}</Text>
               <Text style={{ width: 500 }}>{item.des}</Text>
             </Div>
             <Divider
@@ -74,7 +99,7 @@ export const RenderItem: FC<RenderItemProps> = ({
                 marginVertical: 10,
               }}
             />
-          </>
+          </Div>
         );
       })}
     </Div>
