+++ components/Sidebar/Wrapper.tsx
@@ -14,7 +14,6 @@ const Wrapper = () => {
         <Sidebar onPress={() => router.push("/ui/Button")}>Button</Sidebar>
         <Sidebar onPress={() => router.push("/ui/Card")}>Card</Sidebar>
         <Sidebar onPress={() => router.push("/ui/Checkbox")}>Checkbox</Sidebar>
-        <Sidebar>Card</Sidebar>
         <Sidebar>Div</Sidebar>
         <Sidebar>Divider</Sidebar>
         <Sidebar>Card</Sidebar>
