--- types/Propertics.ts
@@ -1,9 +1,0 @@
-export interface ProperticsProps {
-  name: string;
-  type: string;
-  des: string;
-}
-export interface UsecaseArayObjProps {
-  name: string;
-  code: string;
-}
